requirements.txt
Django==3.2.2
django-stdimage==5.3.0
django-multiselectfield==0.1.12
pytils==0.3
Pillow==8.2.0
djangorestframework==3.12.4
django-filter==2.4.0